Changed how entries are stored in Cassandra
The previous structure used the filename as the key.
Cassandra provides no facility for "discovering" keys which meant the user
needed to know how Scribe named files before accessing any of the entries.

With this scheme, the category is used as the key and the filenames are used
as super-columns names (minus the leading <category>/). Columns are still named
via Time-based UUIDs.

This scheme makes it easy to access the entries for a particular category in sorted order
across multiple "files" with get_slice without needing to use row range operations(which require
using an Ordered Partitioner)